// src/types/entrenadores.ts
export interface Entrenadores {
    entrenador_id?: number
    nombre: string
    apellido: string
    descripcion: string
    correo: string
    experiencia: number
    especialidad: string[]
    precio: number
    estado_entrenador_id?: number
    imagen?: string
    creado_en?: string
    actualizado_en?: string
}   